+++ src/main/java/ru/vk/itmo/chernyshevyaroslav/DiskStorage.java
@@ -3,28 +3,16 @@ package ru.vk.itmo.chernyshevyaroslav;
 import ru.vk.itmo.BaseEntry;
 import ru.vk.itmo.Entry;
 
-import java.io.IOException;
-import java.lang.foreign.Arena;
 import java.lang.foreign.MemorySegment;
 import java.lang.foreign.ValueLayout;
-import java.nio.channels.FileChannel;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.FileAlreadyExistsException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.StandardCopyOption;
-import java.nio.file.StandardOpenOption;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.Iterator;
 import java.util.List;
 import java.util.NoSuchElementException;
-import java.util.concurrent.CopyOnWriteArrayList;
 
 public class DiskStorage {
 
-    protected static final String INDEX_IDX = "index.idx";
-    protected static final String INDEX_TMP = "index.tmp";
     private final List<MemorySegment> segmentList;
 
     public DiskStorage(List<MemorySegment> segmentList) {
@@ -41,138 +29,12 @@ public class DiskStorage {
         }
         iterators.add(firstIterator);
 
-        return new MergeIterator<>(iterators, Comparator.comparing(Entry::key, InMemoryDao::compare)) {
+        return new MergeIterator<>(iterators, Comparator.comparing(Entry::key, ChernyshevDao::compare)) {
             @Override
             protected boolean skip(Entry<MemorySegment> memorySegmentEntry) {
                 return memorySegmentEntry.value() == null;
             }
         };
-    }
-
-    public synchronized static void save(Path storagePath, Iterable<Entry<MemorySegment>> iterable)
-            throws IOException {
-        final Path indexTmp = storagePath.resolve(INDEX_TMP);
-        final Path indexFile = storagePath.resolve(INDEX_IDX);
-
-        try {
-            Files.createFile(indexFile);
-        } catch (FileAlreadyExistsException ignored) {
-            // it is ok, actually it is normal state
-        }
-        List<String> existedFiles = Files.readAllLines(indexFile, StandardCharsets.UTF_8);
-
-        String newFileName = String.valueOf(existedFiles.size());
-
-        long dataSize = 0;
-        long count = 0;
-        for (Entry<MemorySegment> entry : iterable) {
-            dataSize += entry.key().byteSize();
-            MemorySegment value = entry.value();
-            if (value != null) {
-                dataSize += value.byteSize();
-            }
-            count++;
-        }
-        long indexSize = count * 2 * Long.BYTES;
-
-        try (
-                FileChannel fileChannel = FileChannel.open(
-                        storagePath.resolve(newFileName),
-                        StandardOpenOption.WRITE,
-                        StandardOpenOption.READ,
-                        StandardOpenOption.CREATE
-                );
-                Arena writeArena = Arena.ofConfined()
-        ) {
-            MemorySegment fileSegment = fileChannel.map(
-                    FileChannel.MapMode.READ_WRITE,
-                    0,
-                    indexSize + dataSize,
-                    writeArena
-            );
-
-            // index:
-            // |key0_Start|value0_Start|key1_Start|value1_Start|key2_Start|value2_Start|...
-            // key0_Start = data start = end of index
-            long dataOffset = indexSize;
-            int indexOffset = 0;
-            for (Entry<MemorySegment> entry : iterable) {
-                fileSegment.set(ValueLayout.JAVA_LONG_UNALIGNED, indexOffset, dataOffset);
-                dataOffset += entry.key().byteSize();
-                indexOffset += Long.BYTES;
-
-                MemorySegment value = entry.value();
-                if (value == null) {
-                    fileSegment.set(ValueLayout.JAVA_LONG_UNALIGNED, indexOffset, tombstone(dataOffset));
-                } else {
-                    fileSegment.set(ValueLayout.JAVA_LONG_UNALIGNED, indexOffset, dataOffset);
-                    dataOffset += value.byteSize();
-                }
-                indexOffset += Long.BYTES;
-            }
-
-            // data:
-            // |key0|value0|key1|value1|...
-            dataOffset = indexSize;
-            for (Entry<MemorySegment> entry : iterable) {
-                MemorySegment key = entry.key();
-                MemorySegment.copy(key, 0, fileSegment, dataOffset, key.byteSize());
-                dataOffset += key.byteSize();
-
-                MemorySegment value = entry.value();
-                if (value != null) {
-                    MemorySegment.copy(value, 0, fileSegment, dataOffset, value.byteSize());
-                    dataOffset += value.byteSize();
-                }
-            }
-        }
-
-        Files.move(indexFile, indexTmp, StandardCopyOption.ATOMIC_MOVE, StandardCopyOption.REPLACE_EXISTING);
-
-        List<String> list = new ArrayList<>(existedFiles.size() + 1);
-        list.addAll(existedFiles);
-        list.add(newFileName);
-        Files.write(
-                indexFile,
-                list,
-                StandardOpenOption.WRITE,
-                StandardOpenOption.CREATE,
-                StandardOpenOption.TRUNCATE_EXISTING
-        );
-
-        Files.delete(indexTmp);
-    }
-
-    public static List<MemorySegment> loadOrRecover(Path storagePath, Arena arena) throws IOException {
-        Path indexTmp = storagePath.resolve(INDEX_TMP);
-        Path indexFile = storagePath.resolve(INDEX_IDX);
-
-        if (Files.exists(indexTmp)) {
-            Files.move(indexTmp, indexFile, StandardCopyOption.ATOMIC_MOVE, StandardCopyOption.REPLACE_EXISTING);
-        } else {
-            try {
-                Files.createFile(indexFile);
-            } catch (FileAlreadyExistsException ignored) {
-                // it is ok, actually it is normal state
-            }
-        }
-
-        List<String> existedFiles = Files.readAllLines(indexFile, StandardCharsets.UTF_8);
-        List<MemorySegment> result = new CopyOnWriteArrayList<>();
-        for (String fileName : existedFiles) {
-            Path file = storagePath.resolve(fileName);
-            try (FileChannel fileChannel = FileChannel.open(file, StandardOpenOption.READ, StandardOpenOption.WRITE)) {
-                MemorySegment fileSegment = fileChannel.map(
-                        FileChannel.MapMode.READ_WRITE,
-                        0,
-                        Files.size(file),
-                        arena
-                );
-                result.add(fileSegment);
-            }
-        }
-
-        return result;
     }
 
     private static long indexOf(MemorySegment segment, MemorySegment key) {
@@ -278,7 +140,7 @@ public class DiskStorage {
         return segment.byteSize();
     }
 
-    private static long tombstone(long offset) {
+    static long tombstone(long offset) {
         return 1L << 63 | offset;
     }
