+++ src/main/java/ru/vk/itmo/test/chernyshevyaroslav/ChernyshevDaoFactory.java
@@ -3,7 +3,7 @@ package ru.vk.itmo.test.chernyshevyaroslav;
 import ru.vk.itmo.Config;
 import ru.vk.itmo.Dao;
 import ru.vk.itmo.Entry;
-import ru.vk.itmo.chernyshevyaroslav.InMemoryDao;
+import ru.vk.itmo.chernyshevyaroslav.ChernyshevDao;
 import ru.vk.itmo.test.DaoFactory;
 
 import java.io.IOException;
@@ -12,10 +12,10 @@ import java.lang.foreign.ValueLayout;
 import java.nio.charset.StandardCharsets;
 
 @DaoFactory(stage = 5)
-public class InMemoryDaoFactory implements DaoFactory.Factory<MemorySegment, Entry<MemorySegment>> {
+public class ChernyshevDaoFactory implements DaoFactory.Factory<MemorySegment, Entry<MemorySegment>> {
     @Override
     public Dao<MemorySegment, Entry<MemorySegment>> createDao(Config config) throws IOException {
-        return new InMemoryDao(config);
+        return new ChernyshevDao(config);
     }
 
     @Override
